--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "create-react-app": "^5.0.1"
-  }
-}
